+++ docs/changelog.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Place any unreleased changes here, that are subject to release in coming versions :).
 
+* feat: Add structured logs to FastAPI (Uvicorn).
+
 ## 2026-02-23
 
 * feat: Add support for custom COS directories,